Split out former registry manager E2E tests
