+++ e2e/test/iothub/service/DevicesClientE2ETests.cs
@@ -8,9 +8,7 @@ using System.Net;
 using System.Threading.Tasks;
 using Azure;
 using FluentAssertions;
-using Microsoft.Azure.Devices.Client;
 using Microsoft.Azure.Devices.E2ETests.Helpers;
-using Microsoft.Rest;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Azure.Devices.E2ETests.IotHub.Service
@@ -21,9 +19,9 @@ namespace Microsoft.Azure.Devices.E2ETests.IotHub.Service
     [TestClass]
     [TestCategory("E2E")]
     [TestCategory("IoTHub")]
-    public class RegistryE2ETests : E2EMsTestBase
+    public class DevicesClientE2ETests : E2EMsTestBase
     {
-        private readonly string _idPrefix = $"{nameof(RegistryE2ETests)}_";
+        private readonly string _idPrefix = $"{nameof(DevicesClientE2ETests)}_";
 
         private static readonly HashSet<IotHubServiceErrorCode> s_getRetryableStatusCodes = new()
         {
@@ -248,58 +246,6 @@ namespace Microsoft.Azure.Devices.E2ETests.IotHub.Service
 
         [TestMethod]
         [Timeout(TestTimeoutMilliseconds)]
-        public async Task RegistryManager_UpdateTwinsAsync_Works()
-        {
-            // arrange
-
-            var device1 = new Device(_idPrefix + Guid.NewGuid());
-            var device2 = new Device(_idPrefix + Guid.NewGuid());
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-
-            try
-            {
-                await serviceClient.Devices.CreateAsync(device1).ConfigureAwait(false);
-                ClientTwin twin1 = await serviceClient.Twins.GetAsync(device1.Id).ConfigureAwait(false);
-                await serviceClient.Devices.CreateAsync(device2).ConfigureAwait(false);
-                ClientTwin twin2 = await serviceClient.Twins.GetAsync(device2.Id).ConfigureAwait(false);
-
-                // act
-
-                const string expectedProperty = "someNewProperty";
-                const string expectedPropertyValue = "someNewPropertyValue";
-
-                twin1.Properties.Desired[expectedProperty] = expectedPropertyValue;
-                twin2.Properties.Desired[expectedProperty] = expectedPropertyValue;
-
-                BulkRegistryOperationResult result = await serviceClient.Twins
-                    .UpdateAsync(new[] { twin1, twin2 }, false)
-                    .ConfigureAwait(false);
-
-                // assert
-
-                result.IsSuccessful.Should().BeTrue();
-
-                var actualTwin1 = await serviceClient.Twins.GetAsync(device1.Id).ConfigureAwait(false);
-                ((string)actualTwin1.Properties.Desired[expectedProperty]).Should().Be(expectedPropertyValue);
-                var actualTwin2 = await serviceClient.Twins.GetAsync(device2.Id).ConfigureAwait(false);
-                ((string)(actualTwin2.Properties.Desired[expectedProperty])).Should().Be(expectedPropertyValue);
-            }
-            finally
-            {
-                try
-                {
-                    await serviceClient.Devices.DeleteAsync(device1.Id).ConfigureAwait(false);
-                    await serviceClient.Devices.DeleteAsync(device2.Id).ConfigureAwait(false);
-                }
-                catch (Exception ex)
-                {
-                    VerboseTestLogger.WriteLine($"Failed to clean up devices due to {ex}");
-                }
-            }
-        }
-
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
         public async Task DevicesClient_RemoveDevicesAsync_Works()
         {
             // arrange
@@ -360,150 +306,6 @@ namespace Microsoft.Azure.Devices.E2ETests.IotHub.Service
             using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString, options);
             var device = new Device(deviceId);
             await serviceClient.Devices.CreateAsync(device).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
-        public async Task ModulesClient_GetModulesOnDevice()
-        {
-            const int moduleCount = 2;
-            string testDeviceId = $"IdentityLifecycleDevice{Guid.NewGuid()}";
-            string[] testModuleIds = new string[moduleCount];
-            for (int i = 0; i < moduleCount; i++)
-            {
-                testModuleIds[i] = $"IdentityLifecycleModule{i}-{Guid.NewGuid()}";
-            }
-
-            Device device = null;
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-
-            try
-            {
-                // Create a device to house the modules
-                device = await serviceClient.Devices.CreateAsync(new Device(testDeviceId)).ConfigureAwait(false);
-
-                // Create the modules on the device
-                for (int i = 0; i < moduleCount; i++)
-                {
-                    Module createdModule = await serviceClient.Modules.CreateAsync(
-                        new Module(testDeviceId, testModuleIds[i])).ConfigureAwait(false);
-                }
-
-                // Give the hub a moment
-                await Task.Delay(250).ConfigureAwait(false);
-
-                // List the modules on the test device
-                IEnumerable<Module> modulesOnDevice =
-                    await serviceClient.Devices.GetModulesAsync(testDeviceId).ConfigureAwait(false);
-
-                IList<string> moduleIdsOnDevice = modulesOnDevice
-                    .Select(module => module.Id)
-                    .ToList();
-
-                Assert.AreEqual(moduleCount, moduleIdsOnDevice.Count);
-                for (int i = 0; i < moduleCount; i++)
-                {
-                    moduleIdsOnDevice.Should().Contain(testModuleIds[i]);
-                }
-            }
-            finally
-            {
-                await CleanupAsync(serviceClient, testDeviceId).ConfigureAwait(false);
-            }
-        }
-
-        /// <summary>
-        /// Test basic lifecycle of a module.
-        /// This test includes CRUD operations only.
-        /// </summary>
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
-        public async Task ModulesClient_IdentityLifecycle()
-        {
-            string testDeviceId = $"IdentityLifecycleDevice{Guid.NewGuid()}";
-            string testModuleId = $"IdentityLifecycleModule{Guid.NewGuid()}";
-
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-
-            // Create a device to house the module
-            Device device = await serviceClient.Devices.CreateAsync(new Device(testDeviceId)).ConfigureAwait(false);
-
-            try
-            {
-                // Create a module on the device
-                Module createdModule = await serviceClient.Modules.CreateAsync(
-                    new Module(testDeviceId, testModuleId)).ConfigureAwait(false);
-
-                createdModule.DeviceId.Should().Be(testDeviceId);
-                createdModule.Id.Should().Be(testModuleId);
-
-                // Get device
-                // Get the device and compare ETag values (should remain unchanged);
-                Module retrievedModule = null;
-
-                await RetryOperationHelper
-                    .RunWithHubServiceRetryAsync(
-                    async () =>
-                    {
-                        retrievedModule = await serviceClient.Modules.GetAsync(testDeviceId, testModuleId).ConfigureAwait(false);
-                    },
-                    s_retryPolicy)
-                .ConfigureAwait(false);
-
-                retrievedModule.Should().NotBeNull($"When checking for ETag, got null back for GET on module '{testDeviceId}/{testModuleId}'.");
-                retrievedModule.ETag.Should().Be(createdModule.ETag, "ETag value should not have changed between create and get.");
-
-                // Update a module
-                string managedByValue = "SomeChangedValue";
-                retrievedModule.ManagedBy = managedByValue;
-
-                Module updatedModule = await serviceClient.Modules.SetAsync(retrievedModule).ConfigureAwait(false);
-
-                updatedModule.ManagedBy.Should().Be(managedByValue, "Module should have changed its managedBy value");
-
-                // Delete the device
-                // Deleting the device happens in the finally block as cleanup.
-            }
-            finally
-            {
-                await CleanupAsync(serviceClient, testDeviceId).ConfigureAwait(false);
-            }
-        }
-
-        /// <summary>
-        /// Test basic operations of a module's twin.
-        /// </summary>
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
-        public async Task RegistryManager_DeviceTwinLifecycle()
-        {
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-            TestModule module = await TestModule.GetTestModuleAsync(_idPrefix, _idPrefix).ConfigureAwait(false);
-
-            try
-            {
-                // Get the module twin
-                ClientTwin moduleTwin = await serviceClient.Twins.GetAsync(module.DeviceId, module.Id).ConfigureAwait(false);
-
-                moduleTwin.ModuleId.Should().BeEquivalentTo(module.Id, "ModuleId on the Twin should match that of the module identity.");
-
-                // Update device twin
-                string propName = "username";
-                string propValue = "userA";
-                moduleTwin.Properties.Desired[propName] = propValue;
-
-                ClientTwin updatedModuleTwin = await serviceClient.Twins.UpdateAsync(module.DeviceId, module.Id, moduleTwin).ConfigureAwait(false);
-
-                Assert.IsNotNull(updatedModuleTwin.Properties.Desired[propName]);
-                Assert.AreEqual(propValue, (string)updatedModuleTwin.Properties.Desired[propName]);
-
-                // Delete the module
-                // Deleting the module happens in the finally block as cleanup.
-            }
-            finally
-            {
-                await CleanupAsync(serviceClient, module.DeviceId).ConfigureAwait(false);
-            }
         }
 
         [TestMethod]
@@ -628,157 +430,6 @@ namespace Microsoft.Azure.Devices.E2ETests.IotHub.Service
                 {
                     VerboseTestLogger.WriteLine($"Failed to clean up devices due to {ex}");
                 }
-            }
-        }
-
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
-        public async Task ModulesClient_SetModulesETag_Works()
-        {
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-            string deviceId = _idPrefix + Guid.NewGuid();
-            string moduleId = _idPrefix + Guid.NewGuid();
-            var module = new Module(deviceId, moduleId);
-            Device device = await serviceClient.Devices.CreateAsync(new Device(deviceId)).ConfigureAwait(false);
-            module = await serviceClient.Modules.CreateAsync(module).ConfigureAwait(false);
-
-            await RetryOperationHelper
-                .RunWithHubServiceRetryAsync(
-                    async () =>
-                    {
-                        module = await serviceClient.Modules.GetAsync(deviceId, moduleId).ConfigureAwait(false);
-                    },
-                    s_retryPolicy)
-                .ConfigureAwait(false);
-
-            try
-            {
-                ETag oldEtag = module.ETag;
-
-                module.ManagedBy = "test";
-
-                // Update the device once so that the last ETag falls out of date.
-                module = await serviceClient.Modules.SetAsync(module).ConfigureAwait(false);
-
-                // Deliberately set the ETag to an older version to test that the SDK is setting the If-Match
-                // header appropriately when sending the request.
-                module.ETag = oldEtag;
-
-                // set the 'onlyIfUnchanged' flag to true to check that, with an out of date ETag, the request throws a PreconditionFailedException.
-                Func<Task> act = async () => await serviceClient.Modules.SetAsync(module, true).ConfigureAwait(false);
-                var error = await act.Should().ThrowAsync<IotHubServiceException>("Expected test to throw a precondition failed exception since it updated a module with an out of date ETag");
-                error.And.StatusCode.Should().Be(HttpStatusCode.PreconditionFailed);
-                error.And.ErrorCode.Should().Be(IotHubServiceErrorCode.PreconditionFailed);
-                error.And.IsTransient.Should().BeFalse();
-
-                // set the 'onlyIfUnchanged' flag to false to check that, even with an out of date ETag, the request performs without exception.
-                await FluentActions
-                    .Invoking(async () => { module = await serviceClient.Modules.SetAsync(module, false).ConfigureAwait(false); })
-                    .Should()
-                    .NotThrowAsync<IotHubServiceException>("Did not expect test to throw a precondition failed exception since 'onlyIfUnchanged' was set to false");
-
-                // set the 'onlyIfUnchanged' flag to true to check that, with an up-to-date ETag, the request performs without exception.
-                module.ManagedBy = "";
-                device.Status = ClientStatus.Enabled;
-                await FluentActions
-                    .Invoking(async () => { await serviceClient.Modules.SetAsync(module, true).ConfigureAwait(false); })
-                    .Should()
-                    .NotThrowAsync<IotHubServiceException>("Did not expect test to throw a precondition failed exception since 'onlyIfUnchanged' was set to true");
-            }
-            finally
-            {
-                try
-                {
-                    await serviceClient.Modules.DeleteAsync(deviceId, moduleId).ConfigureAwait(false);
-                    await CleanupAsync(serviceClient, deviceId).ConfigureAwait(false);
-                }
-                catch (Exception ex)
-                {
-                    VerboseTestLogger.WriteLine($"Failed to clean up module due to {ex}");
-                }
-            }
-        }
-
-        [TestMethod]
-        [Timeout(TestTimeoutMilliseconds)]
-        public async Task ModulesClient_DeleteModulesETag_Works()
-        {
-            using var serviceClient = new IotHubServiceClient(TestConfiguration.IotHub.ConnectionString);
-            string deviceId = _idPrefix + Guid.NewGuid();
-            string moduleId = _idPrefix + Guid.NewGuid();
-            var module = new Module(deviceId, moduleId);
-            Device device = await serviceClient.Devices.CreateAsync(new Device(deviceId)).ConfigureAwait(false);
-            module = await serviceClient.Modules.CreateAsync(module).ConfigureAwait(false);
-
-            await RetryOperationHelper
-                .RunWithHubServiceRetryAsync(
-                    async () =>
-                    {
-                        module = await serviceClient.Modules.GetAsync(deviceId, moduleId).ConfigureAwait(false);
-                    },
-                    s_retryPolicy)
-                .ConfigureAwait(false);
-
-            try
-            {
-                ETag oldEtag = module.ETag;
-
-                module.ManagedBy = "test";
-
-                // Update the device once so that the last ETag falls out of date.
-                module = await serviceClient.Modules.SetAsync(module).ConfigureAwait(false);
-
-                // Deliberately set the ETag to an older version to test that the SDK is setting the If-Match
-                // header appropriately when sending the request.
-                module.ETag = oldEtag;
-
-                // set the 'onlyIfUnchanged' flag to true to check that, with an out of date ETag, the request throws a PreconditionFailedException.
-                Func<Task> act = async () => await serviceClient.Modules.DeleteAsync(module, true).ConfigureAwait(false);
-                var error = await act.Should().ThrowAsync<IotHubServiceException>("Expected test to throw a precondition failed exception since it updated a module with an out of date ETag");
-                error.And.StatusCode.Should().Be(HttpStatusCode.PreconditionFailed);
-                error.And.ErrorCode.Should().Be(IotHubServiceErrorCode.PreconditionFailed);
-                error.And.IsTransient.Should().BeFalse();
-
-                // set the 'onlyIfUnchanged' flag to false to check that, even with an out of date ETag, the request performs without exception.
-                await FluentActions
-                    .Invoking(async () => { await serviceClient.Modules.DeleteAsync(module, false).ConfigureAwait(false); })
-                    .Should()
-                    .NotThrowAsync<IotHubServiceException>("Did not expect test to throw a precondition failed exception since 'onlyIfUnchanged' was set to false");
-            }
-            finally
-            {
-                try
-                {
-                    await serviceClient.Modules.DeleteAsync(deviceId, moduleId).ConfigureAwait(false);
-                    await CleanupAsync(serviceClient, deviceId).ConfigureAwait(false);
-                }
-                catch (IotHubServiceException ex)
-                    when (ex.StatusCode is HttpStatusCode.NotFound && ex.ErrorCode is IotHubServiceErrorCode.DeviceNotFound)
-                {
-                    // device was already deleted during the normal test flow
-                }
-                catch (Exception ex)
-                {
-                    VerboseTestLogger.WriteLine($"Failed to clean up module due to {ex}");
-                }
-            }
-        }
-
-        private static async Task CleanupAsync(IotHubServiceClient serviceClient, string deviceId)
-        {
-            if (deviceId == null)
-            {
-                return;
-            }
-
-            // cleanup
-            try
-            {
-                await serviceClient.Devices.DeleteAsync(deviceId).ConfigureAwait(false);
-            }
-            catch (Exception ex)
-            {
-                Assert.Fail($"Test clean up of device {deviceId} failed due to {ex}.");
             }
         }
     }
